M-argv-execution step 4: the Windows arm compiles under the real flag set
`heroes run` on Windows answered "internal error: the runtime did not compile"
— the sentence this milestone exists to make honest, and this time it was true.
clang refused runtime/runtime.c there.

One line, in the arm this machine never compiles: hero_str_from_bytes takes an
int64_t and `at` is a size_t, which -Werror=sign-conversion refuses. It sits
inside #if defined(_WIN32), so clang on Darwin does not read it at all.

The lesson is the method rather than the cast. The earlier cross-compile of this
file used -Wall and passed; the project builds with thirteen flags, and the
defect lived in one of the twelve that were missing. A Windows arm is now
compiled against the real flag set before it is pushed — zig cc for
x86_64-windows-gnu and aarch64-windows-gnu, and clang for POSIX, all three clean.

It also shows the repair working from the other side: that error message will
carry clang's own words as soon as the compiler reaches this path with a working
runtime, instead of blaming itself for a machine.

+++ runtime/parts/run.c
@@ -164,7 +164,7 @@ static HeroStr hero_run_win_command_line(void) {
         hero_run_win_quote(hero_run_words[i], line, &at);
     }
     line[at] = '\0';
-    HeroStr built = hero_str_from_bytes(line, at);
+    HeroStr built = hero_str_from_bytes(line, (int64_t)at);
     hero_release(line);
     return built;
 }
